✨ (GEK): Add ExerciseEvaluationStrategy protocol
